Conflicts:
	.gitignore
	plugins/generators/MORSE/org.eclipse.proteus.generators.acceleo.morse/build.acceleo
	plugins/generators/MORSE/org.eclipse.proteus.generators.acceleo.morse/tasks/generate.xml

+++ plugins/robotml/org.eclipse.papyrus.robotml/src/org/eclipse/papyrus/RobotMLLibraries/RobotML_ModelLibrary/RobotML_DataTypes/std_datatypes/impl/Std_datatypesFactoryImpl.java
@@ -558,8 +558,8 @@ public class Std_datatypesFactoryImpl extends EFactoryImpl implements Std_dataty
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public Long createInt64FromString(EDataType eDataType, String initialValue) {
-		return (Long)super.createFromString(eDataType, initialValue);
+	public Double createInt64FromString(EDataType eDataType, String initialValue) {
+		return (Double)super.createFromString(eDataType, initialValue);
 	}
 
 	/**
